JVNGRS-20 Added Sonarqube implementation with root yml edited to ignore test coverage on Application file
This is the commit message #2:

This is the commit message #3:

Please enter the commit message for your changes. Lines starting
with '#' will be ignored, and an empty message aborts the commit.

Date:      Thu Nov 13 16:10:12 2025 +0200

interactive rebase in progress; onto cdc87a5
Last commands done (3 commands done):
   squash 1826b6b # JVNGRS-20 Added SonarQube implementations into yml files and edited the root yml to ignore test coverage on Application files
   squash daf2566 # JVNGRS-20 added property to xml file to ignore test coverage on Application files
No commands remaining.
You are currently rebasing branch 'feature/JVNGRS-20' on 'cdc87a5'.

Changes to be committed:
      modified:   .github/workflows/main.yml
      modified:   .github/workflows/pull_request.yml
      modified:   pom.xml